Replace some of full-include with forward declarations
