+++ src/elona/eobject/eobject.hpp
@@ -1,27 +1,5 @@
 #pragma once
 
+#include "forward.hpp"
 #include "id.hpp"
 #include "ref.hpp"
-
-
-
-namespace elona
-{
-using ObjId = eobject::Id;
-
-
-
-template <typename T>
-using ObjRef = eobject::Ref<T>;
-
-template <typename T>
-using OptionalObjRef = eobject::OptionalRef<T>;
-
-
-
-struct Item;
-
-using ItemRef = ObjRef<Item>;
-using OptionalItemRef = OptionalObjRef<Item>;
-
-} // namespace elona
