+++ src/elona/lua_env/i18n_manager.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../eobject/eobject.hpp"
+#include "../eobject/forward.hpp"
 #include "../optional.hpp"
 #include "lua_submodule.hpp"
