+++ src/elona/position.cpp
@@ -5,6 +5,13 @@
 namespace elona
 {
 
+std::ostream& operator<<(std::ostream& out, const Position& pos)
+{
+    return out << pos.to_string();
+}
+
+
+
 int direction(int x1, int y1, int x2, int y2)
 {
     if (std::abs(x1 - x2) > std::abs(y1 - y2))
